+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Added
+- Clear on Submit option added to the context menu.
+
 ## 1.1.6 [2017-07-26]
 ### Fixed
 - Fixed auto-handling of tickets in order to improve reliability.